requirements.txt
streamlit
PyPDF2
requests
python-dotenv